fix(home-connector): share insecure TLS fetch helper
